requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
gunicorn>=20.1.0
sqlalchemy>=1.4.0
psycopg2-binary>=2.9.0
alembic>=1.7.0
python-dotenv>=0.19.0
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
email-validator>=1.1.3
pydantic>=1.8.0
typing_extensions>=3.10.0

# Testing dependencies
pytest>=6.2.4
httpx>=0.18.0